Added how to play menu
Added SubMenuBase as the base scene for any other menus.
Moved background colour and stars to new scene so it can be used in all menus.
Created HowToMenu scene which contains brief instructions on how to play.
Added buttons for How to play, settings and about to main menu (settings and about only take you to temporary SubMenuBase scene).